% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/booknet.R
\name{booknet}
\alias{booknet}
\title{BookNet}
\usage{
booknet(folder_path, vertex_names, window_length = 15, min_occurences = 0,
  percent_remove = 0)
}
\arguments{
\item{folder_path}{path to a directory which contains the *.txt files (including "names.txt")}

\item{vertex_names}{optional - names for each row/column in the matrix, and vertex labels for `igraph`.
Must be specified in the same order as in names.txt. If missing labels from names.txt will be used instead.}

\item{window_length}{window of words to check for character occurences.}

\item{min_occurences}{weights in the matrix below this number will be set to zero.}

\item{percent_remove}{removes the bottom `percent_remove` percentage of nonzero weights.
If `min_occurences` is specified, those occurences will be removed first, before removing
the bottom percentage specified by `percent_remove`.}
}
\value{
a list, with `$adjacency_matrix` returning the labeled adjacency matrix,
and `$adj_graph` returning an `igraph` graph.
}
\description{
Reads a directory of text files and generates
an adjacency matrix of characters in those text files, with weights
equal to the number of times the relationship appeared.
A connection is made when characters appear within
window_length tokens of each other, delimited by spaces.
The directory must contain one "names.txt" file which
consists of Java regular expressions one each line to search for, for a
given character, ie "Jaime|Kingslayer" would match for
Jaime Lannister in ASOIAF.
}
\examples{
out <- booknet("data/ASOIAF_Folder", min_occurences = 3)
}
